+++ swagger-models/src/test/groovy/com/mangofactory/swagger/models/dto/ApiDescriptionSpec.groovy
@@ -1,7 +1,14 @@
 package com.mangofactory.swagger.models.dto
 
+import com.mangofactory.swagger.models.dto.builder.ApiDescriptionBuilder
+
 class ApiDescriptionSpec extends InternalJsonSerializationSpec {
-  final ApiDescription description = new ApiDescription('p', 'd', [], true)
+  final ApiDescription description = new ApiDescriptionBuilder()
+          .path('p')
+          .description('d')
+          .operations([])
+          .hidden(true)
+          .build()
 
   def "should serialize"() {
     expect: