+++ swagger-springmvc/src/main/java/com/mangofactory/swagger/readers/ApiDescriptionReader.java
@@ -1,11 +1,12 @@
 package com.mangofactory.swagger.readers;
 
 import com.mangofactory.swagger.core.RequestMappingEvaluator;
+import com.mangofactory.swagger.models.dto.ApiDescription;
+import com.mangofactory.swagger.models.dto.Operation;
+import com.mangofactory.swagger.models.dto.builder.ApiDescriptionBuilder;
 import com.mangofactory.swagger.paths.SwaggerPathProvider;
 import com.mangofactory.swagger.readers.operation.RequestMappingReader;
 import com.mangofactory.swagger.scanners.RequestMappingContext;
-import com.mangofactory.swagger.models.dto.ApiDescription;
-import com.mangofactory.swagger.models.dto.Operation;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.mvc.condition.PatternsRequestCondition;
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
@@ -22,8 +23,8 @@ public class ApiDescriptionReader implements Command<RequestMappingContext> {
   private final RequestMappingEvaluator requestMappingEvaluator;
 
   public ApiDescriptionReader(SwaggerPathProvider pathProvider,
-      Collection<RequestMappingReader> customAnnotationReaders,
-      RequestMappingEvaluator requestMappingEvaluator) {
+                              Collection<RequestMappingReader> customAnnotationReaders,
+                              RequestMappingEvaluator requestMappingEvaluator) {
     this.swaggerPathProvider = pathProvider;
     this.customAnnotationReaders = customAnnotationReaders;
     this.requestMappingEvaluator = requestMappingEvaluator;
@@ -46,7 +47,13 @@ public class ApiDescriptionReader implements Command<RequestMappingContext> {
         ApiOperationReader apiOperationReader = new ApiOperationReader(customAnnotationReaders);
         apiOperationReader.execute(context);
         List<Operation> operations = (List<Operation>) context.get("operations");
-        apiDescriptionList.add(new ApiDescription(path, methodName, operations, false));
+        apiDescriptionList.add(
+                new ApiDescriptionBuilder()
+                        .path(path)
+                        .description(methodName)
+                        .operations(operations)
+                        .hidden(false)
+                        .build());
       }
     }
     context.put("apiDescriptionList", apiDescriptionList);